create separate func for isAdmin validation

+++ controllers/categoryController.js
@@ -1,46 +1,48 @@
 import Category from "../models/category.js";
 
-export function postCategory(req, res) {
-  if (req.user == null) {
-    res.status(401).json({ msg: "Please login to create a category" });
-    return;
-  }
-  if (req.user.type != "customer") {
-    res.status(403).json({
-      msg: "only admin can create a category",
-    });
-    return;
+function isAdmin(req, res) {
+  if (!req.user) {
+    res.status(401).json({ msg: "Please login" });
+    return false;
   }
 
-  const newCategory = new Category(req.body);
-  newCategory
-    .save()
-    .then((result) =>
-      res.json({ msg: "Category created successfully", result: result })
-    )
-    .catch((err) =>
-      res.json({
-        msg: "Failed to create category",
-        error: err,
-      })
-    );
+  if (req.user.type !== "admin") {
+    res.status(403).json({ msg: "Only admin can use this function" });
+    return false;
+  }
+  return true;
+}
+
+export function postCategory(req, res) {
+  if (isAdmin(req, res)) {
+    const newCategory = new Category(req.body);
+    newCategory
+      .save()
+      .then((result) =>
+        res.json({ msg: "Category created successfully", result: result })
+      )
+      .catch((err) =>
+        res.status(500).json({
+          msg: "Failed to create category",
+          error: err,
+        })
+      );
+  }
 }
 
 export function getCategory(req, res) {
-  Category.find().then(
-    (result)=>{
+  Category.find()
+    .then((result) => {
       res.json({
         category: result,
       });
-    }
-  ).catch(
-    (err)=>{
+    })
+    .catch((err) => {
       res.json({
         msg: "Failed to fetch categories",
         error: err,
       });
-    }
-  )
+    });
 }
 
 export function getCategoryByName(req, res) {
@@ -60,17 +62,19 @@ export function getCategoryByName(req, res) {
 }
 
 export function deleteCategoryByName(req, res) {
-  const name = req.params.name;
-  Category.deleteOne({ name: name })
-    .then((result) => {
-      res.json({
-        result:result
+  if (isAdmin(req, res)) {
+    const name = req.params.name;
+    Category.deleteOne({ name: name })
+      .then((result) => {
+        res.json({
+          result: result,
+        });
+      })
+      .catch((err) => {
+        res.json({
+          msg: "Failed to fetch category by name",
+          error: err,
+        });
       });
-    })
-    .catch((err) => {
-      res.json({
-        msg: "Failed to fetch category by name",
-        error: err,
-      });
-    });
+  }
 }
